Securing scripts

+++ uninstall.php
@@ -1,5 +1,7 @@
 <?php
 
+if(!defined("ABSPATH")) die;
+
 delete_option('socials_accounts');
 
 ?>
